Combine all type of tasks in one class Task.
