+++ src/test/java/TaskTest.java
@@ -1,8 +1,7 @@
-import java.util.Arrays;
-import java.util.List;
-
 import org.junit.*;
 import static org.junit.Assert.*;
+import org.junit.Rule;
+import java.util.List;
 
 public class TaskTest {
 
@@ -10,73 +9,62 @@ public class TaskTest {
   public DatabaseRule database = new DatabaseRule();
 
   @Test
-  public void all_emptyAtFirst() {
-    assertEquals(Task.all().size(), 0);
-  }
-
-  @Test
-  public void save_savesTaskIntoDatabase() {
-    User user = new User("PM", "Nathan", "nathan@pmam.com");
-    user.save();
-    Task Task = new Task("Title", user.getId(), "In Progress", "This isn't a real Task", 1, 2);
-    Task.save();
-    assertEquals("Title", Task.find(Task.getId()).getTitle());
-  }
-
-  @Test
-  public void Task_deleteWorksProperly() {
-    User user = new User("PM", "Nathan", "nathan@pmam.com");
-    user.save();
-    Task Task = new Task("Title", user.getId(), "In Progress", "This isn't a real Task", 1, user.getId());
-    Task.save();
-    Task.delete();
-    assertEquals(Task.all().size(), 0);
-  }
-
-  @Test
-  public void equals_returnsTrueIfSame() {
-    User user = new User("PM", "Nathan", "nathan@pmam.com");
-    user.save();
-    Task firstTask = new Task("Title", 1, "In Progress", "This isn't a real Task", 1, 2);
-    firstTask.save();
-    Task secondTask = new Task("Title", 1, "In Progress", "This isn't a real Task", 1, 2);
-    secondTask.save();
-    assertTrue(firstTask.equals(secondTask));
-  }
-
-  @Test
-  public void Task_updateTitleAndDescriptionWorksProperly() {
-    User user = new User("PM", "Nathan", "nathan@pmam.com");
-    user.save();
-    Task Task = new Task("Title", 1, "In Progress", "This isn't a real Task", 1, 2);
-    Task.save();
-    Task.updateTitleAndDescription("New title", "haaaaaands");
-    assertEquals(Task.getTitle(), "New title");
-    assertEquals(Task.find(Task.getId()).getTitle(), "New title");
-  }
-
-  @Test
-  public void Task_updateStatus() {
-    User user = new User("PM", "Nathan", "nathan@pmam.com");
-    user.save();
-    Task Task = new Task("Title", 1, "In Progress", "This isn't a real Task", 1, 2);
-    Task.save();
-    Task.updateStatus("Completed");
-    assertEquals(Task.getStatus(), "Completed");
-    assertEquals(Task.find(Task.getId()).getStatus(), "Completed");
-  }
-
-  @Test
-  public void Task_updateImplementer() {
-    User user = new User("PM", "Nathan", "nathan@pmam.com");
-    user.save();
-    User newUser = new User("PM", "Andrey", "nathan@pmam.com");
+  public void task_createsNewTask() {
+    User newUser = new User("PM", "Nathan", "nathan@pmam.com");
     newUser.save();
-    Task Task = new Task("Title", user.getId(), "In Progress", "This isn't a real Task", 1, user.getId());
-    Task.save();
-    Task.updateImplementer(newUser.getId());
-    assertEquals(Task.getImplementerId(), newUser.getId());
-    assertEquals(Task.find(Task.getId()).getImplementerId(), newUser.getId());
+    Task newTask = new Task("Some title", newUser.getId(), 1, "small description", 1, newUser.getId());
+    newTask.save();
+    Task savedTask = Task.find(newTask.getId());
+    assertTrue(savedTask.equals(newTask));
+    assertEquals(newTask.getTitle(), "Some title");
+    assertEquals(newTask.getStatus(), 1);
+    assertEquals(newTask.getDescription(), "small description");
+    assertEquals(newTask.getTypeTask(), 1);
+    assertEquals(newTask.getCreatorId(), newUser.getId());
+  }
+
+
+  @Test
+  public void task_Updated() {
+    User newUser = new User("PM", "Nathan", "nathan@pmam.com");
+    newUser.save();
+    Task newTask = new Task("Some title", newUser.getId(), 1, "small description", 1, newUser.getId());
+    newTask.save();
+    newTask.update("Some other title", "Large description");
+    assertEquals(newTask.getTitle(), "Some other title");
+    assertEquals(newTask.getDescription(), "Large description");
+  }
+
+  @Test
+  public void task_UpdatesStatus() {
+    User newUser = new User("PM", "Nathan", "nathan@pmam.com");
+    newUser.save();
+    Task newTask = new Task("Some title", newUser.getId(), 1, "small description", 1, newUser.getId());
+    newTask.save();
+    newTask.updateStatus(2);
+    assertEquals(newTask.getStatus(), 2);
+  }
+
+  @Test
+  public void task_UpdatesImplementor() {
+    User newUser = new User("PM", "Nathan", "nathan@pmam.com");
+    newUser.save();
+    Task newTask = new Task("Some title", newUser.getId(), 1, "small description", 1, newUser.getId());
+    newTask.save();
+    User secondUser = new User("PM", "Illia", "illia@pmam.com");
+    secondUser.save();
+    newTask.updateImplementor(secondUser.getId());
+    assertEquals(newTask.getImplementorId(), secondUser.getId());
+  }
+
+  @Test
+  public void task_deleteTaskWithAnyTypes() {
+    User newUser = new User("PM", "Nathan", "nathan@pmam.com");
+    newUser.save();
+    Task newTask = new Task("Some title", newUser.getId(), 1, "small description", 1, newUser.getId());
+    newTask.save();
+    newTask.delete();
+    assertEquals(0, Task.all(1).size());
   }
 
 }
